Introduction_To_Programming_with_Python/final_project/requirements.txt
beautifulsoup4==4.12.3
certifi==2023.11.17
cfgv==3.4.0
charset-normalizer==3.3.2
distlib==0.3.8
exceptiongroup==1.2.0
filelock==3.13.1
identify==2.5.33
idna==3.6
iniconfig==2.0.0
nodeenv==1.8.0
packaging==23.2
platformdirs==4.1.0
pluggy==1.4.0
pre-commit==3.6.0
pytest==8.0.0
python-dotenv==1.0.1
PyYAML==6.0.1
requests==2.31.0
soupsieve==2.5
tomli==2.0.1
urllib3==2.1.0
virtualenv==20.25.0
